+++ generators/entity/index.js
@@ -28,6 +28,7 @@ const statistics = require('../statistics');
 const { isReservedClassName, isReservedTableName } = require('../../jdl/jhipster/reserved-keywords');
 const { entityDefaultConfig } = require('../generator-defaults');
 const { prepareEntityForTemplates } = require('../../utils/entity');
+const { prepareFieldForTemplates } = require('../../utils/field');
 
 /* constants used throughout */
 const SUPPORTED_VALIDATION_RULES = constants.SUPPORTED_VALIDATION_RULES;
@@ -486,12 +487,12 @@ class EntityGenerator extends BaseBlueprintGenerator {
 
                 // Load in-memory data for fields
                 this.context.fields.forEach(field => {
-                    this._prepareFieldForTemplates(field);
+                    prepareFieldForTemplates(entity, field, generator);
                 });
 
                 // Load in-memory data for relationships
                 this.context.relationships.forEach(relationship => {
-                    this._prepareRelationshipForTemplates(relationship);
+                    prepareRelationshipForTemplates(relationship);
                 });
             },
 
@@ -790,82 +791,6 @@ class EntityGenerator extends BaseBlueprintGenerator {
                 BASE_TEMPLATE_DATA
             )
         );
-    }
-
-    _prepareFieldForTemplates(field, entity = this.context) {
-        const fieldOptions = field.options || {};
-        _.defaults(field, {
-            fieldNameCapitalized: _.upperFirst(field.fieldName),
-            fieldNameUnderscored: _.snakeCase(field.fieldName),
-            fieldNameHumanized: fieldOptions.fieldNameHumanized || _.startCase(field.fieldName),
-        });
-        const fieldType = field.fieldType;
-
-        field.fieldIsEnum = this.fieldIsEnum(fieldType);
-
-        if (field.fieldNameAsDatabaseColumn === undefined) {
-            const fieldNameUnderscored = _.snakeCase(field.fieldName);
-            const jhiFieldNamePrefix = this.getColumnName(entity.jhiPrefix);
-            if (isReservedTableName(fieldNameUnderscored, entity.prodDatabaseType)) {
-                if (!jhiFieldNamePrefix) {
-                    this.warning(
-                        `The field name '${fieldNameUnderscored}' is regarded as a reserved keyword, but you have defined an empty jhiPrefix. This might lead to a non-working application.`
-                    );
-                    field.fieldNameAsDatabaseColumn = fieldNameUnderscored;
-                } else {
-                    field.fieldNameAsDatabaseColumn = `${jhiFieldNamePrefix}_${fieldNameUnderscored}`;
-                }
-            } else {
-                field.fieldNameAsDatabaseColumn = fieldNameUnderscored;
-            }
-            field.columnName = field.fieldNameAsDatabaseColumn;
-        }
-
-        if (field.fieldInJavaBeanMethod === undefined) {
-            // Handle the specific case when the second letter is capitalized
-            // See http://stackoverflow.com/questions/2948083/naming-convention-for-getters-setters-in-java
-            if (field.fieldName.length > 1) {
-                const firstLetter = field.fieldName.charAt(0);
-                const secondLetter = field.fieldName.charAt(1);
-                if (firstLetter === firstLetter.toLowerCase() && secondLetter === secondLetter.toUpperCase()) {
-                    field.fieldInJavaBeanMethod = firstLetter.toLowerCase() + field.fieldName.slice(1);
-                } else {
-                    field.fieldInJavaBeanMethod = _.upperFirst(field.fieldName);
-                }
-            } else {
-                field.fieldInJavaBeanMethod = _.upperFirst(field.fieldName);
-            }
-        }
-
-        if (field.fieldValidateRulesPatternJava === undefined) {
-            field.fieldValidateRulesPatternJava = field.fieldValidateRulesPattern
-                ? field.fieldValidateRulesPattern.replace(/\\/g, '\\\\').replace(/"/g, '\\"')
-                : field.fieldValidateRulesPattern;
-        }
-
-        if (field.fieldValidateRulesPatternAngular === undefined) {
-            field.fieldValidateRulesPatternAngular = field.fieldValidateRulesPattern
-                ? field.fieldValidateRulesPattern.replace(/"/g, '&#34;')
-                : field.fieldValidateRulesPattern;
-        }
-
-        if (field.fieldValidateRulesPatternReact === undefined) {
-            field.fieldValidateRulesPatternReact = field.fieldValidateRulesPattern
-                ? field.fieldValidateRulesPattern.replace(/'/g, "\\'")
-                : field.fieldValidateRulesPattern;
-        }
-
-        field.fieldValidate = Array.isArray(field.fieldValidateRules) && field.fieldValidateRules.length >= 1;
-        field.nullable = !(field.fieldValidate === true && field.fieldValidateRules.includes('required'));
-        field.unique = field.fieldValidate === true && field.fieldValidateRules.includes('unique');
-        if (field.unique) {
-            field.uniqueConstraintName = this.getUXConstraintName(entity.entityTableName, field.columnName, entity.prodDatabaseType);
-        }
-        if (field.fieldValidate === true && field.fieldValidateRules.includes('maxlength')) {
-            field.maxlength = field.fieldValidateRulesMaxlength || 255;
-        }
-
-        return field;
     }
 
     _prepareRelationshipForTemplates(relationship, entity = this.context) {
